fix elliptical errors to sum precision matrices, add tests
This only affects the case when NWAY is run with asymmetric errors.
Previously, the inverse variances (precision) and precision matrices
were erroneously multiplied.
Luckily, this was detected by Dong-Woo Kim in the form of strange
Bayes factors favoring the most distant sources. The origin of
this is that the distances became negative, because the
covariance and precision matrices are not semi-positive definite anymore.

This commit makes the simple change of adding the precision matrices.
This is consistent with the formulas of Pineau+18, although
their normalisation factors are different. In essence, this is bridging
the Pineau+ and Budavari+ approaches yet again.